feat(spell-service,spell-dtos)!: Refactor Spell Service API + Update Installation Spell to the new API [fixes NET-643] (#222)
Breaking changes in Spell Service API.

API changes:
- The value fields of the `*Value` structures are renamed to `value`
(instead of `str`, `num`, `v1_str`, ...)
- AIR script store/load functions are now: `get_script`/`set_script`.
- Removed `get_script_code_from_env`
- Removed location API and `set_relay`
- Renamed `list_remove_value` -> `list_remove_string` (dunno I didn't
named it like this in the first place)

Behavioural fixes:
- When trying to access a value of a wrong type, return 'absent = true`.
Previously, we agreed to return an error, but due to some sqlite
problems[0] it's _easier_ just to mark values like non-existent.
   Example: 
   ```
  1> call spell set_u32 ["a", 2]
   result: {
  "error": "",
  "success": true
   }
   2> call spell get_string ["a"]
  result: {
  "absent": true,
  "error": "",
  "success": true,
  "value": ""
  }
   ```
- **BUT** the above doesn't work for lists. Both a list and a value can
exist by the **same key** at the **same time**.
The list functions don't affect the values and vice versa. It's much
_easier_ to implement this way.
  Example of KV
  ```
  sqlite> select * from kv;
  +-----+--------+-----+------------+
  | key | string | u32 | list_order |
  +-----+--------+-----+------------+
  | a   |        | 2   | -1         |  <--- U32
  | a   | b      |     | 0          |  <--- List
  +-----+--------+-----+------------+
  ```
- Supported storage of empty strings. 
The behavior was inconsistent due to wasm problems with empty
strings[0]. I implemented some workarounds, so we can store and read
empty strings.
  

[0] On reading and writing an empty string from/to sqlite, the value
transforms to NULL. Also, in rust code when we tried to do
`Some(empty_string)` it transformed to `None` due to internal
representation of nulls and empty values. That led to storage behavior.